Add a keyring that contains an attacker's pub/sec key.  Use this to maliciously sign a download file.
In get_files(), check the signatures of all downloaded files if there is both
a .asc and a non-.asc file available.  Note that if no .asc or non-.asc file
was downloaded, but should have been, then we would have gotten a 404 at
download time, and the same FileNotFoundError result.