LP: #1421434 - When the udm downloader creates a group download and waits for its finished signal, the reactor must only respond to signals on the object_path for the group download it created.  It turns out that in some scenarios, multiple udm reactors can be active at a time and if they don't filter on object_path, a reactor whose group download is not complete may terminate prematurely.  (This was the root cause of LP: #1413265, which I've closed as invalid.)
Add another test for 100% udm code coverage.

Simplify the _print() helper in the udm downloader and remove a test that was
depending on the previous behavior (and only existed for code coverage).

Simply the tox.ini so that two environment variables which are not dependent
on the test environment, can be shared.