Add and include session and instance in user_agent
This adds and includes session and instance in user_agent, this way we
can make and calculate better device stats

sessionId is a random id that is created on boot, this way we can
calculate how often the devices reboot and with this we can calculate
the stability of our builds.

instanceId is hashed and splitted in half android/device serial number,
this number is not in any way traceable since it's both hashed an
splitted in half, this way this way it's impossible to reverse. With
this id we can calulate how many devices is using ubports, this can also
help see how many new users, and also see how many users we loose