Add RO-RTC support to system-image
Most newer Android devices have a read-only hardware clock. This means
that we need to store an offset value from that clock to the current
time and reload it at system startup to retain a correct time.
Unfortunately, our recovery doesn't do that. Instead, it just keeps
the default time (which is usually no later than sometime in 1975).
This causes a small artifact when querying for system-image's last
updated metric, the date is shown as sometime in the 1970s rather than
when you actually updated.

This commit makes system-image read the offset from timekeeper
(ubports/timekeeper) and use that when it determines the time of the
last update. Fixes ubports/ubuntu-touch#685